Protect UDPSocket as well
