+++ src/relay/utils.rs
@@ -1,17 +1,9 @@
 use std::{
     future::Future,
-    io::{self, Error, ErrorKind},
-    io::prelude::*,
-    net::{SocketAddr, TcpStream},
+    io::{self, Error},
     time::Duration,
-    os::unix::io::{AsRawFd, RawFd},
-    os::unix::net::UnixStream,
 };
 
-use log::{info};
-
-use net2::*;
-use sendfd::*;
 use tokio::time;
 
 pub async fn try_timeout<T, E, F>(fut: F, timeout: Option<Duration>) -> io::Result<T>
@@ -49,50 +41,4 @@ pub fn set_nofile(_nofile: u64) -> io::Result<()> {
     //
     // Windows' limit of opening files is the size of HANDLE (32-bits), so it is unlimited
     Ok(())
-}
-
-#[cfg(target_os="android")]
-pub fn protect(protect_path: &Option<String>, fd: RawFd) -> io::Result<()> {
-    // ignore if protect_path is not specified
-    let path = match protect_path {
-        Some(path) => path,
-        None => return Ok(()),
-    };
-
-    // it's safe to use blocking socket here
-    let mut stream = UnixStream::connect(path)?;
-    stream.set_read_timeout(Some(Duration::new(1, 0))).expect("couldn't set read timeout");
-    stream.set_write_timeout(Some(Duration::new(1, 0))).expect("couldn't set write timeout");
-
-    // send fds
-    let dummy: [u8; 1] = [1];
-    let fds: [RawFd; 1] = [fd];
-    stream.send_with_fd(&dummy, &fds)?;
-
-    // receive the return value
-    let mut response = [0; 1];
-    stream.read(&mut response)?;
-
-    if response[0] == 0xFF {
-        return Err(Error::new(ErrorKind::Other, "protect() failed"));
-    }
-
-    Ok(())
-}
-
-#[cfg(not(target_os="android"))]
-pub fn protect(protect_path: &Option<String>, fd: RawFd) -> io::Result<()> {
-    Ok(())
-}
-
-// create a new TCP stream
-pub fn new_tcp_stream(protect_path: &Option<String>, saddr: &SocketAddr) -> io::Result<TcpStream> {
-    let builder = match saddr {
-        SocketAddr::V4(_) => TcpBuilder::new_v4()?,
-        SocketAddr::V6(_) => TcpBuilder::new_v6()?,
-    };
-
-    protect(protect_path, builder.as_raw_fd())?;
-
-    builder.to_tcp_stream()
 }
